+++ src/irdevice/light.ts
@@ -56,13 +56,17 @@ export class Light {
 
   async OnSet(value: CharacteristicValue): Promise<void> {
     this.debugLog(`${this.device.remoteType}: ${this.accessory.displayName} On: ${value}`);
-    this.On = value;
-    this.accessory.context.On = this.On;
-    if (this.On) {
+    if (value) {
       await this.pushLightOnChanges();
     } else {
       await this.pushLightOffChanges();
     }
+    /**
+     * pushLightOnChanges and pushLightOffChanges above assume they are measuring the state of the accessory BEFORE
+     * they are updated, so we are only updating the accessory state after calling the above.
+     */
+    this.On = value;
+    this.accessory.context.On = this.On;
   }
 
   /**
@@ -183,7 +187,7 @@ export class Light {
     }
   }
 
-  async commandType() {
+  async commandType(): Promise<string> {
     let commandType: string;
     if (this.device.customize) {
       commandType = 'customize';
@@ -193,7 +197,7 @@ export class Light {
     return commandType;
   }
 
-  async commandOn() {
+  async commandOn(): Promise<string> {
     let command: string;
     if (this.device.customize && this.device.customOn) {
       command = this.device.customOn;
@@ -203,10 +207,10 @@ export class Light {
     return command;
   }
 
-  async commandOff() {
+  async commandOff(): Promise<string> {
     let command: string;
-    if (this.device.customize && this.device.customOn) {
-      command = this.device.customOn;
+    if (this.device.customize && this.device.customOff) {
+      command = this.device.customOff;
     } else {
       command = 'turnOff';
     }