+++ src/irdevice/tv.ts
@@ -206,6 +206,12 @@ export class TV {
       } else {
         await this.pushTvOnChanges();
       }
+      /**
+       * pushTvOnChanges and pushTvOffChanges above assume they are measuring the state of the accessory BEFORE
+       * they are updated, so we are only updating the accessory state after calling the above.
+       */
+      this.Active = value;
+      this.accessory.context.Active = this.Active;
     }
   }
 
@@ -396,7 +402,7 @@ export class TV {
     }
   }
 
-  async commandType() {
+  async commandType(): Promise<string> {
     let commandType: string;
     if (this.device.customize) {
       commandType = 'customize';
@@ -406,7 +412,7 @@ export class TV {
     return commandType;
   }
 
-  async commandOn() {
+  async commandOn(): Promise<string> {
     let command: string;
     if (this.device.customize && this.device.customOn) {
       command = this.device.customOn;
@@ -416,10 +422,10 @@ export class TV {
     return command;
   }
 
-  async commandOff() {
+  async commandOff(): Promise<string> {
     let command: string;
-    if (this.device.customize && this.device.customOn) {
-      command = this.device.customOn;
+    if (this.device.customize && this.device.customOff) {
+      command = this.device.customOff;
     } else {
       command = 'turnOff';
     }